# api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
...
  /estate/{id}/tree:
    post:
      summary: Stores tree data in a given estate.
      operationId: addTreeToEstate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the estate where the tree will be added.
      requestBody:
        description: Tree details containing plot coordinates (x, y) and height.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreeRequest"
      responses:
        '201':
          description: Tree added successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeResponse"
        '400':
          description: Invalid value or format received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Estate not found.
...
components:
  schemas:
...
    TreeRequest:
      type: object
      required:
        - x
        - y
        - height
      properties:
        x:
          type: integer
          minimum: 1
          maximum: 50000
          description: X coordinate of the tree's plot location
        y:
          type: integer
          minimum: 1
          maximum: 50000
          description: Y coordinate of the tree's plot location
        height:
          type: integer
          minimum: 1
          maximum: 30
          description: Height of the tree in meters (must be between 1 and 30)
    TreeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          description: UUID of the created tree
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string